+++ singletons/core_initializer.gd
@@ -224,7 +224,7 @@ var _procedural_classes: Dictionary = IVGlobal.procedural_classes
 
 
 func _enter_tree() -> void:
-	IVFiles.init_from_config(self, IVGlobal.config, "core_initializer")
+	IVFiles.init_from_config(self, IVGlobal.ivoyager_config, "core_initializer")
 
 
 func _ready() -> void:
@@ -301,7 +301,7 @@ func _instantiate_preinitializers() -> void:
 
 func _do_presets_and_plugin_mods() -> void:
 	# TODO: We might add class presets here
-	if !IVGlobal.tree_saver_enabled:
+	if !IVPluginUtils.is_plugin_enabled("ivoyager_tree_saver"):
 		program_nodes.erase(&"SaveManager")
 		gui_nodes.erase(&"SaveDialog")
 		gui_nodes.erase(&"LoadDialog")